# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "test Posiflora",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/notifications": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/taskController.CreateOneRequestTask"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/baseModel.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/taskController.CreateOneRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/taskController.CreateOneResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/baseModel.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "baseModel.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                },
                "is_error": {
                    "type": "boolean"
                }
            }
        },
        "taskController.CreateOneRequest": {
            "type": "object",
            "properties": {
                "task": {
                    "$ref": "#/definitions/taskController.CreateOneRequestTask"
                }
            }
        },
        "taskController.CreateOneRequestTask": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "email_sender": {
                    "type": "string"
                },
                "recipient_email": {
                    "type": "string"
                }
            }
        },
        "taskController.CreateOneResponse": {
            "type": "object",
            "properties": {
                "task": {
                    "$ref": "#/definitions/taskController.Task"
                }
            }
        },
        "taskController.Task": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}